refactor: shared shell utils
