+++ shared/package.json
@@ -21,6 +21,7 @@
     "@monokle/validation": "0.7.1",
     "@open-policy-agent/opa-wasm": "1.8.0",
     "@reduxjs/toolkit": "1.9.0",
+    "electron": "^21.2.3",
     "electron-store": "^8.1.0",
     "lodash": "4.17.21",
     "nodemon": "2.0.20",
