+++ src/utils/shell.ts
@@ -1,39 +1,13 @@
-import {app, shell} from 'electron';
-
-import * as os from 'os';
-
-import {WINDOW_HELP_LINK} from '@monokle-desktop/shared/constants/telemetry';
-
-import {trackEvent} from './telemetry';
+import {shell} from 'electron';
 
 export function showItemInFolder(fullPath: string) {
   shell.showItemInFolder(fullPath);
-}
-
-export function openGitHub() {
-  trackEvent(WINDOW_HELP_LINK, {linkID: 'github'});
-  shell.openExternal('https://github.com/kubeshop/monokle');
-}
-
-export function openDiscord() {
-  trackEvent(WINDOW_HELP_LINK, {linkID: 'discord'});
-  shell.openExternal('https://discord.gg/kMJxmuYTMu');
 }
 
 export function openUrlInExternalBrowser(url?: string) {
   if (url) {
     shell.openExternal(url);
   }
-}
-
-export function openDocumentation() {
-  trackEvent(WINDOW_HELP_LINK, {linkID: 'documentation'});
-  shell.openExternal(`https://kubeshop.github.io/monokle?os=${os.type}`);
-}
-
-export function openLogs() {
-  trackEvent(WINDOW_HELP_LINK, {linkID: 'logs'});
-  shell.showItemInFolder(app.getPath('logs'));
 }
 
 export function openUniqueObjectNameTopic() {
